pyproject.toml
[tool.poetry]
name = "discord-nativ-reddit-vids"
version = "0.1.0"
description = ""
authors = ["Nico Krätschmer <kraetschmerni@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.3.2"
yt-dlp = "^2023.10.13"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
